adds hot reload
