+++ .babelrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   presets: ['react-app'],
   plugins: [
+    'react-hot-loader/babel',
     '@babel/plugin-proposal-class-properties',
     '@babel/plugin-proposal-optional-chaining',
   ],
