+++ package.json
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "5.12.0",
+    "@hot-loader/react-dom": "17.0.1",
     "@react-pdf/renderer": "2.0.0",
     "@reduxjs/toolkit": "1.5.1",
     "as-table": "1.0.55",
@@ -81,7 +82,6 @@
     "@electron-forge/maker-rpm": "6.0.0-beta.54",
     "@electron-forge/maker-squirrel": "6.0.0-beta.54",
     "@electron-forge/plugin-webpack": "6.0.0-beta.54",
-    "@hot-loader/react-dom": "16.11.0",
     "@marshallofsound/webpack-asset-relocator-loader": "0.5.0",
     "@types/classnames": "2.2.11",
     "@types/dateformat": "3.0.1",
