+++ src/main.ts
@@ -14,9 +14,6 @@ import sourceMapSupport from 'source-map-support'
 
 import MenuBuilder from './menu'
 
-declare const MAIN_WINDOW_WEBPACK_ENTRY: string
-declare const MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY: string
-
 // Enable dev tools
 if (!app.isPackaged) {
   app.whenReady().then(() => {
