+++ webpack.renderer.config.js
@@ -37,5 +37,8 @@ module.exports = {
   plugins: plugins,
   resolve: {
     extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
+    alias: {
+      'react-dom': '@hot-loader/react-dom',
+    },
   },
 }
